Upvote and Downvote Buttons Work
While the buttons work for sending upvotes and downvotes, they do not currently change color to tell the user that they have been clicked. We will
implement a backend endpoint that tells a user their upvote/downvote status so that each user can have peristent state. Then we will adjust the
front-end code to change the color of the arrows based on this state.